Move all Dagger DI modules under di/modules directory since Kotlin does not have "package private" access privilege.
