+++ app/src/main/java/com/github/brianspace/moviebrowser/di/components/AppComponent.kt
@@ -18,16 +18,16 @@ package com.github.brianspace.moviebrowser.di.components
 
 import com.github.brianspace.moviebrowser.MovieBrowserApplication
 import com.github.brianspace.moviebrowser.di.modules.AppModule
-import com.github.brianspace.moviebrowser.di.modules.FavoriteMovieListFragmentModule
-import com.github.brianspace.moviebrowser.di.modules.MainActivityModule
-import com.github.brianspace.moviebrowser.di.modules.MovieDetailsActivityModule
-import com.github.brianspace.moviebrowser.di.modules.NowPlayingMovieListFragmentModule
-import com.github.brianspace.moviebrowser.di.modules.SettingsFragmentModule
-import com.github.brianspace.moviebrowser.models.ModelsModule
-import com.github.brianspace.moviebrowser.repository.local.LocalRepositoryModule
-import com.github.brianspace.moviebrowser.repository.util.DirUtilModule
-import com.github.brianspace.moviebrowser.repository.web.MovieDbServiceModule
-import com.github.brianspace.moviebrowser.viewmodels.ViewModelsModule
+import com.github.brianspace.moviebrowser.di.modules.models.ModelsModule
+import com.github.brianspace.moviebrowser.di.modules.repository.DirUtilModule
+import com.github.brianspace.moviebrowser.di.modules.repository.LocalRepositoryModule
+import com.github.brianspace.moviebrowser.di.modules.repository.MovieDbServiceModule
+import com.github.brianspace.moviebrowser.di.modules.ui.FavoriteMovieListFragmentModule
+import com.github.brianspace.moviebrowser.di.modules.ui.MainActivityModule
+import com.github.brianspace.moviebrowser.di.modules.ui.MovieDetailsActivityModule
+import com.github.brianspace.moviebrowser.di.modules.ui.NowPlayingMovieListFragmentModule
+import com.github.brianspace.moviebrowser.di.modules.ui.SettingsFragmentModule
+import com.github.brianspace.moviebrowser.di.modules.viewmodels.ViewModelsModule
 import dagger.Component
 import dagger.android.AndroidInjectionModule
 import dagger.android.AndroidInjector
