+++ app/src/main/java/com/github/brianspace/moviebrowser/di/modules/ui/FavoriteMovieListFragmentModule.kt
@@ -1,4 +1,4 @@
-package com.github.brianspace.moviebrowser.di.modules
+package com.github.brianspace.moviebrowser.di.modules.ui
 
 import com.github.brianspace.moviebrowser.ui.fragment.FavoriteMovieListFragment
 import dagger.Module
